refactor: simplify files
