+++ jolt-core/src/poly/commitment/bmmtv.rs
@@ -1,7 +1,7 @@
 #![deny(warnings, unused, nonstandard_style)]
 #![allow(clippy::type_complexity, clippy::upper_case_acronyms)]
 
-pub mod commitments;
+pub mod afgho;
 pub mod gipa;
 pub mod inner_products;
 pub mod tipa;
