+++ jolt-core/src/poly/commitment/bmmtv/afgho.rs
@@ -1,9 +1,22 @@
+//! This crate defines what is a Double Homomorphic Commitment and
+//! exports some implementations of it like Afgho, Identity and Pedersen
+
 use std::marker::PhantomData;
 
-use ark_ec::pairing::{Pairing, PairingOutput};
+use ark_ec::{
+    pairing::{Pairing, PairingOutput},
+    CurveGroup,
+};
 use ark_std::rand::Rng;
 
-use super::{super::inner_products::PairingInnerProduct, random_generators, Error};
+use super::inner_products::PairingInnerProduct;
+
+pub type Error = anyhow::Error;
+
+/// Helpers for generator commitment keys used by Pedersen and AFGHO16
+pub fn random_generators<R: Rng, G: CurveGroup>(rng: &mut R, num: usize) -> Vec<G> {
+    (0..num).map(|_| G::rand(rng)).collect()
+}
 
 /// Afgho commitment is simply an inner pairing product commitment
 #[derive(Clone)]
