--- jolt-core/src/poly/commitment/bmmtv/tipa/structured_scalar_message.rs
@@ -1,206 +1,0 @@
-use std::marker::PhantomData;
-
-use ark_ec::pairing::{Pairing, PairingOutput};
-use ark_ff::One;
-use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
-use ark_std::{cfg_iter, end_timer, start_timer};
-
-use super::super::{
-    gipa::Gipa,
-    tipa::{prove_commitment_key_kzg_opening, verify_commitment_key_g2_kzg_opening},
-    Error,
-};
-use crate::{
-    field::JoltField,
-    msm::Icicle,
-    poly::commitment::{
-        bmmtv::{
-            commitments::afgho16::AfghoCommitment, gipa::CommitmentSteps,
-            inner_products::MultiexponentiationInnerProduct,
-        },
-        kzg::KZGVerifierKey,
-    },
-    utils::transcript::Transcript,
-};
-
-/// Multiexponentiation with known field vector
-///
-/// In the MIPPk protocol a prover demonstrates knowledge of [A] ∈ [G1] such
-/// that A commits to pairing commitment T under *v* and U = A^b for a public vector [b] ∈ [F].
-pub struct MippK<P, Transcript> {
-    _pair: PhantomData<P>,
-    _transcript: PhantomData<Transcript>,
-}
-
-/// Proof of [`MippK`]
-#[derive(CanonicalSerialize, CanonicalDeserialize, Clone)]
-pub struct MippKProof<P>
-where
-    P: Pairing,
-{
-    commitment_steps: CommitmentSteps<P>,
-    final_message: (P::G1, P::ScalarField),
-    final_ck: P::G2,
-    final_ck_proof: P::G2,
-}
-
-impl<P, ProofTranscript> MippK<P, ProofTranscript>
-where
-    P: Pairing,
-    P::G1: Icicle,
-    P::ScalarField: JoltField,
-    ProofTranscript: Transcript,
-{
-    pub fn prove(
-        h_beta_powers: &[P::G2],
-        values: (&[P::G1], &[P::ScalarField]),
-        commitment_key: &[P::G2],
-        transcript: &mut ProofTranscript,
-    ) -> Result<MippKProof<P>, Error> {
-        // Run GIPA
-        let gipa = start_timer!(|| "GIPA");
-        let proof = Gipa::<P, ProofTranscript>::prove(
-            (values.0.to_vec(), values.1.to_vec()),
-            commitment_key.to_vec(),
-            transcript,
-        )?;
-        end_timer!(gipa);
-
-        // Prove final commitment key is wellformed
-        let ck_kzg = start_timer!(|| "Prove commitment key");
-        let ck_a_final = proof.final_commitment_param;
-        let transcript_inverse = cfg_iter!(proof.scalar_transcript)
-            .map(|x| JoltField::inverse(x).unwrap())
-            .collect::<Vec<_>>();
-
-        // KZG challenge point
-        transcript.append_serializable(&ck_a_final);
-        let c = transcript.challenge_scalar();
-
-        // Complete KZG proof
-        let ck_a_kzg_opening = prove_commitment_key_kzg_opening(
-            h_beta_powers,
-            &transcript_inverse,
-            P::ScalarField::one(), // r_shift = one, why?
-            c,
-        )?;
-        end_timer!(ck_kzg);
-
-        Ok(MippKProof {
-            final_message: proof.final_message,
-            commitment_steps: proof.commitment_steps,
-            final_ck: ck_a_final,
-            final_ck_proof: ck_a_kzg_opening,
-        })
-    }
-
-    pub fn verify(
-        v_srs: &KZGVerifierKey<P>,
-        com: (PairingOutput<P>, P::G1),
-        scalar_b: P::ScalarField,
-        proof: &MippKProof<P>,
-        transcript: &mut ProofTranscript,
-    ) -> Result<bool, Error> {
-        let (base_com, gipa_transcript) =
-            Gipa::<P, ProofTranscript>::verify(com, &proof.commitment_steps, transcript)?;
-        let transcript_inverse = cfg_iter!(gipa_transcript)
-            .map(|x| JoltField::inverse(x).unwrap())
-            .collect::<Vec<_>>();
-
-        // KZG challenge point
-        transcript.append_serializable(&proof.final_ck);
-        let c = transcript.challenge_scalar();
-
-        // Check commitment key
-        let ck_a_valid = verify_commitment_key_g2_kzg_opening(
-            v_srs,
-            proof.final_ck,
-            proof.final_ck_proof,
-            &transcript_inverse,
-            P::ScalarField::one(),
-            c,
-        )?;
-
-        // Compute final scalar
-        let mut power_2_b = scalar_b;
-        let mut product_form = Vec::new();
-        for x in gipa_transcript.iter() {
-            product_form
-                .push(<P::ScalarField>::one() + (JoltField::inverse(x).unwrap() * power_2_b));
-            power_2_b *= power_2_b;
-        }
-        let b_base = cfg_iter!(product_form).product::<P::ScalarField>();
-
-        // Verify base inner product commitment
-        let (com_a, com_t) = base_com;
-        let a_base = vec![proof.final_message.0];
-        let t_base = MultiexponentiationInnerProduct::inner_product(&a_base, &[b_base])?;
-
-        let same_ip_commit = t_base == com_t;
-        let base_valid =
-            AfghoCommitment::verify(&[proof.final_ck], &a_base, &com_a)? && same_ip_commit;
-
-        Ok(ck_a_valid && base_valid)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::sync::Arc;
-
-    use ark_bn254::Bn254;
-    use ark_std::{
-        rand::{rngs::StdRng, SeedableRng},
-        UniformRand,
-    };
-
-    use super::{
-        super::super::commitments::{afgho16::AfghoCommitment, random_generators},
-        *,
-    };
-    use crate::{
-        poly::commitment::{bmmtv::tipa::Field, kzg::SRS},
-        utils::transcript::KeccakTranscript,
-    };
-
-    type BnAfghoG1 = AfghoCommitment<Bn254>;
-    type BnScalarField = <Bn254 as Pairing>::ScalarField;
-
-    const TEST_SIZE: usize = 8;
-
-    fn structured_scalar_power<F: Field>(num: usize, s: &F) -> Vec<F> {
-        let mut powers = vec![F::one()];
-        for i in 1..num {
-            powers.push(powers[i - 1] * s);
-        }
-        powers
-    }
-
-    #[test]
-    fn tipa_ssm_multiexponentiation_inner_product_test() {
-        type MultiExpTipa = MippK<Bn254, KeccakTranscript>;
-
-        let mut rng = StdRng::seed_from_u64(0u64);
-        let srs = SRS::setup(&mut rng, 2 * (TEST_SIZE - 1), 2 * (TEST_SIZE - 1));
-
-        let ck_a = srs.get_commitment_keys();
-        let powers_len = srs.g1_powers.len();
-        let (p_srs, v_srs) = SRS::trim(Arc::new(srs), powers_len - 1);
-
-        let m_a = random_generators(&mut rng, TEST_SIZE);
-        let b = BnScalarField::rand(&mut rng);
-        let m_b = structured_scalar_power(TEST_SIZE, &b);
-        let com_a = BnAfghoG1::commit(&ck_a, &m_a).unwrap();
-        let com_t = MultiexponentiationInnerProduct::inner_product(&m_a, &m_b).unwrap();
-
-        let mut transcript = KeccakTranscript::new(b"TipaTest");
-
-        let proof =
-            MultiExpTipa::prove(&p_srs.h_beta_powers(), (&m_a, &m_b), &ck_a, &mut transcript)
-                .unwrap();
-
-        let mut transcript = KeccakTranscript::new(b"TipaTest");
-
-        assert!(MultiExpTipa::verify(&v_srs, (com_a, com_t), b, &proof, &mut transcript,).unwrap());
-    }
-}
